Online Pump control code for esp32
The provided code facilitates communication between an microcontroller (ESP32) and the Blynk cloud, enabling continuous reading of the value on pin V0 from the cloud. It then provide feedback to the actuator which is a the relay module connected with micocontroller based on changes to the V0 value on the Blynk cloud.